feat: добавление загрузки фото; улучшение кода по 12.12 (часть 1)
