+++ js/gallery.js
@@ -1,7 +1,32 @@
 import {showBigPicture} from './big-picture.js';
-import {renderGallery} from './render-gallery.js';
 
-const picturesContainer = document.querySelector('.pictures');
+const picturesContainer = document.querySelector('.pictures-container');
+const pictureTemplate = document.querySelector('#picture')
+  .content
+  .querySelector('.picture');
+
+const createPicture = (picture) => {
+  const clonePicture = pictureTemplate.cloneNode(true);
+
+  clonePicture.querySelector('.picture__likes').textContent = picture.likes;
+  clonePicture.querySelector('.picture__comments').textContent = picture.comments.length;
+  clonePicture.querySelector('.picture__img').src = picture.url;
+  clonePicture.querySelector('.picture__img').alt = picture.description;
+  clonePicture.dataset.pictureId = picture.id;
+
+  return clonePicture;
+};
+
+export const renderGallery = (pictures) => {
+  picturesContainer.innerHTML = '';
+  const picturesFragment = document.createDocumentFragment();
+
+  pictures.forEach((picture) => {
+    picturesFragment.append(createPicture(picture));
+  });
+
+  picturesContainer.append(picturesFragment);
+};
 
 export const renderPicturesGallery = (pictures) => {
   picturesContainer.addEventListener('click', (evt) => {
@@ -19,6 +44,4 @@ export const renderPicturesGallery = (pictures) => {
       showBigPicture(photo);
     }
   });
-
-  renderGallery(pictures);
 };
