+++ js/picture-sort.js
@@ -1,10 +1,7 @@
-import {debounce, sortByRandomly} from './utils.js';
-import {renderGallery} from './render-gallery.js';
+import {debounce} from './utils.js';
+import {renderGallery} from './gallery.js';
 
 const MAX_PICTURE_COUNT = 10;
-
-const sortingSection = document.querySelector('.img-filters');
-const sortingForm = sortingSection.querySelector('.img-filters__form');
 
 const SORTING_TYPE = {
   DEFAULT: 'filter-default',
@@ -12,8 +9,13 @@ const SORTING_TYPE = {
   DISCUSSED: 'filter-discussed'
 };
 
+const sortingSection = document.querySelector('.img-filters');
+const sortingForm = sortingSection.querySelector('.img-filters__form');
+
 const sortByComments = (pictureA, pictureB) =>
   pictureB.comments.length - pictureA.comments.length;
+
+const returnRandomNumber = () => Math.random() - 0.5;
 
 let originalPictures = [];
 
@@ -22,7 +24,7 @@ const sortPicturesByType = (sortingType) => {
 
   switch (sortingType) {
     case SORTING_TYPE.RANDOM:
-      return sortByRandomly(ordinaryPictures).slice(0, MAX_PICTURE_COUNT);
+      return ordinaryPictures.sort(returnRandomNumber).slice(0, MAX_PICTURE_COUNT);
     case SORTING_TYPE.DISCUSSED:
       return ordinaryPictures.sort(sortByComments);
     default:
@@ -31,10 +33,12 @@ const sortPicturesByType = (sortingType) => {
 };
 
 const onSortingClick = debounce((evt) => {
-  if (evt.target.closest('.img-filters__button')) {
-    sortingForm.querySelector('.img-filters__button--active').classList.remove('img-filters__button--active');
-    evt.target.classList.add('img-filters__button--active');
+  if (!evt.target.closest('.img-filters__button')) {
+    return;
   }
+
+  sortingForm.querySelector('.img-filters__button--active').classList.remove('img-filters__button--active');
+  evt.target.classList.add('img-filters__button--active');
 
   const sortedPictures = sortPicturesByType(evt.target.id);
   renderGallery(sortedPictures);
